+++ shuup/admin/modules/categories/views/__init__.py
@@ -4,7 +4,8 @@
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
+from .copy import CategoryCopyVisibilityView
 from .edit import CategoryEditView
 from .list import CategoryListView
 
-__all__ = ["CategoryEditView", "CategoryListView"]
+__all__ = ["CategoryEditView", "CategoryListView", "CategoryCopyVisibilityView"]